requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
reportlab
gunicorn
uvicorn
channels
channels-redis
dj-database-url
daphne
whitenoise
cloudinary
django-cloudinary-storage